Moved some files around and quick renames
